Add exemplars into metrics.proto
This makes metrics.proto a bit more OpenMetrics-like. Note that this
is not an attempt to make metrics.proto a full and valid
representation of OpenMetrics. It's merely added here so that v1 of
prometheus/client_golang (which uses metrics.proto as the internal
data model) can expose OpenMetrics text format.

Signed-off-by: beorn7 <beorn@grafana.com>