Flatten the buckets of native histograms
This is slightly more efficient on the wire, and it matches the way
the corresponding Go types are structured (`histogram.Histogram` and
`histogram.FloatHistogram`). This follow the change for the
remote-write protobuf in
https://github.com/prometheus/prometheus/pull/11011 .

This commit also moves away from the "sparse histogram" naming in lieu
of the preferred name "native histogram".

Note that this is deliberately an incompatible change of the proto
spec. Keeping compatibility would cause some hassle but not much gain
because we haven't published the proto spec in any release yet and
always marked it as experimental. Compatibility to the released proto
spec (without native histograms) is kept, of course.

Signed-off-by: beorn7 <beorn@grafana.com>